requirements.txt
streamlit
openai
langchain-openai
python-dotenv
langchain